refactor: move navbar to header slot
